refactor: move usecase to upper level
